Refactor: Split AgentServer.cs
